Rename: auth-client import 경로 수정
