+++ src/features/profile/update-nickname/api/use-update-nickname.ts
@@ -1,8 +1,7 @@
 import { groupQuery } from "@/entities/group";
 import { profileQuery } from "@/entities/profile";
 import { scheduleQuery } from "@/entities/schedule";
-import { patch } from "@/features/auth/login";
-import { errorToast, successToast } from "@/shared/lib";
+import { errorToast, patch, successToast } from "@/shared/lib";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 
 export const useUpdateNickname = () => {
