+++ src/features/respond-schedule/api/use-update-schedule-response.ts
@@ -1,6 +1,5 @@
 import { scheduleQuery } from "@/entities/schedule/api/schedule.query";
-import { patch } from "@/features/auth/login";
-import { errorToast, successToast } from "@/shared/lib";
+import { errorToast, patch, successToast } from "@/shared/lib";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { RESPOND_SCHEDULE_MESSAGES } from "../model/constants";
